Create Array+Identifiable.swift file and move the firstIndex(matching:) function to the new extension file;
