+++ MemorizeCS193p/Array+Identifiable.swift
@@ -1,0 +1,19 @@
+//
+//  Array+Identifiable.swift
+//  MemorizeCS193p
+//
+//  Created by William Spanfelner on 14/10/2021.
+//
+
+import Foundation
+
+extension Array where Element: Identifiable {
+    func firstIndex(matching: Element) -> Int {
+        for index in 0 ..< self.count {
+            if self[index].id == matching.id {
+                return index
+            }
+        }
+        return 0 // FIXME: - BOGUS!!
+    }
+}
