+++ MemorizeCS193p/Grid.swift
@@ -35,14 +35,3 @@ struct Grid<Item, ItemView>: View where Item: Identifiable, ItemView: View {
             .position(layout.location(ofItemAt: index))
     }
 }
-
-extension Array where Element: Identifiable {
-    func firstIndex(matching: Element) -> Int {
-        for index in 0 ..< self.count {
-            if self[index].id == matching.id {
-                return index
-            }
-        }
-        return 0 // FIXME: - BOGUS!!
-    }
-}
